docs-2978: move link-check skip lists to one config file
Move the crawler skip list and skip-origins list out of __tests__/crawler.test.js,
and the default skip and ignore lists out of src/utils/linkChecker.js, into a new
shared module src/utils/link-check-config.js. Both files import from it.

Pure move, no behavior change: entry counts are unchanged (crawler skip 152,
default skip 34, default ignore 3, skip-origins 6).
